+++ src/main/java/com/project/goalchallenge/global/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.project.goalchallenge.global.config.exception;
+package com.project.goalchallenge.global.exception;
 
 import lombok.Getter;
